rename .resource to .request
