+++ api/src/resources/Accounts.ts
@@ -1,5 +1,5 @@
 //// codegen:start {preset: barrel, include: ../Blog/*.ts, export: { as: 'PascalCase' }}
-export { GetMe } from "../Accounts/GetMe.resource.js"
+export { GetMe } from "../Accounts/GetMe.request.js"
 //// codegen:end
 
 // codegen:start {preset: meta, sourcePrefix: src/resources/}
