+++ api/src/resources/Users.ts
@@ -1,4 +1,4 @@
-export { IndexUsers } from "../Accounts/IndexUsers.resource.js"
+export { IndexUsers } from "../Accounts/IndexUsers.request.js"
 
 // codegen:start {preset: meta, sourcePrefix: src/resources/}
 export const meta = { moduleName: "Users" } as const
